+++ src/main/java/roomescape/reservation/controller/ReservationUserViewController.java
@@ -1,10 +1,10 @@
-package roomescape.controller;
+package roomescape.reservation.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class UserViewController {
+public class ReservationUserViewController {
 
     @GetMapping("admin")
     public String home() {
